util: refactor PrioritizedTaskExecutor
