+++ packages/trie/src/util/index.ts
@@ -1,4 +1,3 @@
 export * from './encoding.js'
 export * from './genesisState.js'
-export * from './tasks.js'
 export * from './walkController.js'
