+++ packages/util/src/index.ts
@@ -65,4 +65,5 @@ export * from './lock.js'
 export * from './mapDB.js'
 export * from './provider.js'
 export * from './requests.js'
+export * from './tasks.js'
 export * from './verkle.js'
